Patch custom x-axis function to fix case where values are not integers. Added a couple of arguments and reorganised files.
