+++ src/dfeqa/__init__.py
@@ -1,3 +1,4 @@
 from .db import load_census
-from .education_data import valid_name_regex, valid_upn, year_group
+from .data_validation import valid_name_regex, valid_upn
 from .summaries import summarise_str_lengths, fd, freqchart, parse_text
+from .data_transformation import year_group
